:truck: Split gflow_wrapper
